MOVE setOptions to layout module REMOVE merge from initial state MOVE initial state of layout & ui to layout module ADD merge of initial, restored to layout module
