+++ lib/ui/src/core/context.js
@@ -61,7 +61,7 @@ export class Provider extends Component {
       storyId,
     };
 
-    this.state = {};
+    this.state = store.getInitialState();
 
     this.modules = [
       initChannel,
@@ -75,7 +75,7 @@ export class Provider extends Component {
     ].map(initModule => initModule(apiData));
 
     // Create our initial state by combining the initial state of all modules, then overlaying any saved state
-    const state = store.getInitialState(getInitialState(...this.modules.map(m => m.state)));
+    const state = getInitialState(...this.modules.map(m => m.state));
 
     // Get our API by combining the APIs exported by each module
     const combo = Object.assign({ navigate }, ...this.modules.map(m => m.api));
