+++ lib/ui/src/core/store.js
@@ -3,8 +3,6 @@
 import { localStorage, sessionStorage } from 'global';
 import { parse, stringify } from 'telejson';
 
-import merge from '../libs/merge';
-
 export const STORAGE_KEY = '@storybook/ui/store';
 
 function get(storage) {
@@ -32,11 +30,11 @@ export default class Store {
 
   // The assumption is that this will be called once, to initialize the React state
   // when the module is instanciated
-  getInitialState(base = {}) {
+  getInitialState() {
     // We don't only merge at the very top level (the same way as React setState)
     // when you set keys, so it makes sense to do the same in combining the two storage modes
     // Really, you shouldn't store the same key in both places
-    return merge(base, { ...get(localStorage), ...get(sessionStorage) });
+    return { ...get(localStorage), ...get(sessionStorage) };
   }
 
   getState() {
